LayoutTests/webcl/tests/webcl-test-arg-types-const.html

<html>
<head>
<title>WebCL WebCLKernelArgumentTypes constants tests.</title>
<link rel="stylesheet" href="../resources/js-test-style.css"/>
<script type="text/javascript" src="../resources/js-test-pre.js"></script>
<script type="text/javascript">

function runTest() {
    shouldBeTrue("WebCLKernelArgumentTypes.CHAR == 0");
    shouldBeTrue("WebCLKernelArgumentTypes.UCHAR == 1");
    shouldBeTrue("WebCLKernelArgumentTypes.SHORT == 2");
    shouldBeTrue("WebCLKernelArgumentTypes.USHORT == 3");
    shouldBeTrue("WebCLKernelArgumentTypes.INT == 4");
    shouldBeTrue("WebCLKernelArgumentTypes.UINT == 5");
    shouldBeTrue("WebCLKernelArgumentTypes.LONG == 6");
    shouldBeTrue("WebCLKernelArgumentTypes.ULONG == 7");
    shouldBeTrue("WebCLKernelArgumentTypes.FLOAT == 8");
    shouldBeTrue("WebCLKernelArgumentTypes.HALF == 9");
    shouldBeTrue("WebCLKernelArgumentTypes.DOUBLE == 10");
    shouldBeTrue("WebCLKernelArgumentTypes.VEC2 == 0x0200");
    shouldBeTrue("WebCLKernelArgumentTypes.VEC3 == 0x0300");
    shouldBeTrue("WebCLKernelArgumentTypes.VEC4 == 0x0400");
    shouldBeTrue("WebCLKernelArgumentTypes.VEC8 == 0x0800");
    shouldBeTrue("WebCLKernelArgumentTypes.VEC16 == 0x1000");
    shouldBeTrue("WebCLKernelArgumentTypes.LOCAL_MEMORY_SIZE == 255");
}
</script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
    description(document.title);
    debug("Test all constants defined in WebCLKernelArgumentTypes.idl");
    runTest();
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
